Improved build process
Will now clone the repository to a temporary folder.
Like this, the build process will not influence the project in
the working directory.
This prepares it to be built in a CI/CD environment.
Added new property `files_dir`, which determines where
to look for files, and where they will be cloned to.
The cloning will clone the current project using the filesystem,
if new `repoUrl` property is not specified.
The branch will be auto-determined to be the current branch,
if `branch` is not specified.